# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Identity Service
  version: 2.0.0
  description: >
    Central identity and access management service.
    Handles member registration, authentication (email/password + social OAuth),
    legacy member migration, session management, token verification,
    organisational hierarchy, group management, and permissions.
servers:
  - url: http://localhost:3000
    description: Local development
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:      { type: string }
            message:   { type: string }
            requestId: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field:   { type: string }
                  message: { type: string }
                required: [message]
          required: [code, message, requestId, details]
      required: [error]
    TokenPair:
      type: object
      properties:
        accessToken:  { type: string }
        refreshToken: { type: string }
      required: [accessToken, refreshToken]
...
    Installation:
      type: object
      properties:
        id:         { type: string }
        name:       { type: string }
        code:       { type: string }
        timezone:   { type: string }
        is_active:  { type: boolean }
        created_at: { type: string, format: date-time }
      required: [id, name, code, timezone, is_active, created_at]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT issued by the identity service.
        Obtain via /v1/auth/register, /v1/auth/login, or OAuth flows.
paths:
...
  /v1/auth/register/installation:
    get:
      summary: Resolve installation code to UUID
      operationId: resolveInstallation
      tags: [Auth]
      description: Public endpoint. Called by the frontend before rendering the registration form.
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Installation resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Installation' }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/auth/register:
    post:
      summary: Register a new member
      operationId: register
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName, installationId]
              properties:
                email:           { type: string, format: email }
                password:        { type: string, minLength: 8 }
                firstName:       { type: string, minLength: 1 }
                lastName:        { type: string, minLength: 1 }
                installationId:  { type: string }
                nationId:        { type: string }
                tribeId:         { type: string }
                workforceTeamId: { type: string }
                phone:           { type: string }
                address:         { type: string }
                dateOfBirth:     { type: string, format: date }
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/TokenPair' }
        '400': { description: Validation error or invalid installation, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: Email already taken, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
  /v1/auth/migrate/password:
    post:
      summary: Set password for legacy member
      operationId: migratePassword
      tags: [Migration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [migrationToken, password]
              properties:
                migrationToken: { type: string }
                password:       { type: string, minLength: 8 }
      responses:
        '200':
          description: Migrated and authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/TokenPair' }
        '400': { description: Invalid token, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/auth/migrate/oauth/{provider}:
    get:
      summary: Start OAuth migration
      operationId: migrateOAuthStart
      tags: [Migration]
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string, enum: [google, apple, facebook] } }
        - { name: migrationToken, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        '302': { description: Redirect to provider }
  /v1/auth/migrate/oauth/{provider}/callback:
    get:
      summary: OAuth migration callback
      operationId: migrateOAuthCallback
      tags: [Migration]
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string, enum: [google, apple, facebook] } }
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        '302': { description: Redirect to frontend with tokens }
  /v1/auth/oauth/{provider}:
    get:
      summary: Initiate social OAuth login
      operationId: oauthStart
      tags: [OAuth]
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string, enum: [google, apple, facebook] } }
        - { name: state, in: query, required: true, schema: { type: string } }
        - { name: installationCode, in: query, required: false, schema: { type: string }, description: Installation code from registration URL }
      responses:
        '302': { description: Redirect to provider }
  /v1/auth/oauth/{provider}/callback:
    get:
      summary: OAuth callback
      operationId: oauthCallback
      tags: [OAuth]
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string, enum: [google, apple, facebook] } }
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        '302': { description: Redirect to frontend with tokens }
  /v1/auth/password/forgot:
    post:
      summary: Request password reset email
      operationId: forgotPassword
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200': { description: Reset email sent if account exists }
  /v1/auth/password/reset:
    post:
      summary: Reset password
      operationId: resetPassword
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:    { type: string }
                password: { type: string, minLength: 8 }
      responses:
        '204': { description: Password reset }
        '400': { description: Invalid token, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/auth/session/refresh:
    post:
      summary: Refresh access token
      operationId: refreshSession
      tags: [Session]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      accessToken: { type: string }
        '401': { description: Invalid token, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/auth/session/logout:
    post:
      summary: Logout
      operationId: logout
      tags: [Session]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '204': { description: Logged out }
...